+++ scripts/sensors/wind_speed.py
@@ -19,7 +19,7 @@ interval = helper_functions.getCron()[0]
 test = True
 if len(sys.argv) > 1:
 	rest = int(sys.argv[1])
-	iterations = (interval*60/rest)-1
+	iterations = int((interval*60/rest)-1)
 elif os.isatty(sys.stdin.fileno()) or test_toggle == "true":
 	rest = 10
 	iterations = (interval*6)-1
